Simplify the users repo
